--- custom-metronome/src/components/Login/Login.tsx
@@ -1,9 +1,0 @@
-import React, { Component } from 'react'
-
-export default class Login extends Component {
-  render() {
-    return (
-      <div>Login</div>
-    )
-  }
-}
